+++ lib/invoke.js
@@ -1,7 +1,6 @@
 const request = require('request-promise');
 const os = require('os');
-const stream = require('stream');
-const { BUILD_URL, getSDKVersion, parseResponse } = require('../utils/utils');
+const { BUILD_URL, getSDKVersion, parseResponse, parseData } = require('../utils/utils');
 
 module.exports = (key, service, action, data, isCLI) => {
   const headers = {
@@ -10,18 +9,8 @@ module.exports = (key, service, action, data, isCLI) => {
     'X-Build-Meta-OS': `${os.type().toLowerCase()}@${os.release()}`,
   };
 
-  const files = {};
-  const filelessData = {};
-  for (const param in data) {
-    if (data[param] instanceof stream.Readable || Buffer.isBuffer(data[param])) {
-      files[param] = data[param];
-    } else {
-      filelessData[param] = data[param];
-    }
-  }
-
   return request.post(`${BUILD_URL}/run/${prefixName(service)}/${action}`, {
-    formData: Object.assign({}, files, { data: JSON.stringify(filelessData) }),
+    formData: parseData(data),
     resolveWithFullResponse: true,
     headers,
     auth: { user: key },
@@ -30,7 +19,7 @@ module.exports = (key, service, action, data, isCLI) => {
     return parseResponse(response);
   }).catch((err) => {
     if (err.response && err.response.headers) checkDeprecated(err.response);
-    return Promise.reject(err);
+    return Promise.reject(err.error);
   });
 };
